Broke up FilterBar into SearchRecipes component and FilterByCategories component. Added comments for each component
